+++ src/ckanapi_harvesters/ckan_api/ckan_api_4_readwrite.py
@@ -74,7 +74,7 @@ class CkanApiReadWriteParams(CkanApiPolicyParams):
 class CkanApiReadWrite(CkanApiPolicy):
     """
     CKAN Database API interface to CKAN server with helper functions using pandas DataFrames.
-    This class implements requests to write data to the CKAN server resources / DataStores.
+    This extension implements requests to write data to the CKAN server resources / DataStores.
     """
 
     def __init__(self, url:str=None, *, proxies:Union[str,dict,ProxyConfig]=None,