+++ src/ckanapi_harvesters/ckan_api/ckan_api_7_sysadmin.py
@@ -6,15 +6,16 @@
 from typing import List, Union
 from contextlib import contextmanager
 
+from ckanapi_harvesters.auxiliary.ckan_action import CkanNotFoundError
 from ckanapi_harvesters.auxiliary.proxy_config import ProxyConfig
 from ckanapi_harvesters.policies.data_format_policy import CkanPackageDataFormatPolicy
 from ckanapi_harvesters.harvesters.data_cleaner.data_cleaner_abc import CkanDataCleanerABC
 
 from ckanapi_harvesters.auxiliary.ckan_map import CkanMap
 from ckanapi_harvesters.auxiliary.ckan_api_key import CkanApiKey
-from ckanapi_harvesters.auxiliary.ckan_model import CkanApiTokenInfo
+from ckanapi_harvesters.auxiliary.ckan_model import CkanApiTokenInfo, CkanUserInfo, CkanGroupInfo, CkanState
 from ckanapi_harvesters.auxiliary.ckan_auxiliary import RequestType
-from ckanapi_harvesters.auxiliary.ckan_errors import (ArgumentError)
+from ckanapi_harvesters.auxiliary.ckan_errors import (ArgumentError, DuplicateNameError)
 from ckanapi_harvesters.ckan_api.ckan_api_5_manage import CkanApiExtendedParams
 from ckanapi_harvesters.ckan_api.ckan_api_6_user_access import CkanApiUserAccess
 
@@ -27,7 +28,7 @@ class CkanApiSysadminParams(CkanApiExtendedParams):
 class CkanApiSysadmin(CkanApiUserAccess):
     """
     CKAN Database API interface to CKAN server with helper functions using pandas DataFrames.
-    This class implements methods only available to sysadmins... under development.
+    This extension implements methods which bear no direct relation with data manipulation.
     """
 
     def __init__(self, url:str=None, *, proxies:Union[str,dict,ProxyConfig]=None,
@@ -128,4 +129,97 @@ class CkanApiSysadmin(CkanApiUserAccess):
             return True
         else:
             raise response.default_error(self)
+
+    #% ---------- User management
+    def _api_user_create(self, *, name: str, email: str, password: str, fullname: str=None,
+                         about: str = None, image_url: str = None, plugin_extras: dict=None, params:dict=None) -> CkanUserInfo:
+        """
+        Call to API user_create
+        Creates a new user (requires sysadmin user)
+        """
+        if params is None:
+            params = {}
+        params["name"] = name
+        params["email"] = email
+        params["password"] = password
+        if fullname is not None:
+            params["fullname"] = fullname
+        if about is not None:
+            params["about"] = about
+        if image_url is not None:
+            params["image_url"] = image_url
+        if plugin_extras is not None:
+            params["plugin_extras"] = plugin_extras
+        response = self._api_action_request(f"user_create", method=RequestType.Post, json=params)
+        if response.success:
+            user_info = CkanUserInfo.from_dict(response.result)
+            # update map
+            self.map._update_user_info(user_info.copy())
+            return user_info
+        else:
+            raise response.default_error(self)
+
+    def user_create(self, *, name: str, email: str, password: str, fullname: str = None,
+                    about: str = None, image_url: str = None, plugin_extras: dict = None,
+                    params:dict = None, error_exists: bool = True) -> CkanUserInfo:
+        """
+        Call to API user_create, checking at first if the user does not exist.
+        """
+        user_found = True
+        try:
+            _ = self.user_show(name)
+        except CkanNotFoundError:
+            user_found = False
+        if user_found and error_exists:
+            raise DuplicateNameError("user", name)
+        return self._api_user_create(name=name, email=email, password=password, fullname=fullname,
+                                     about=about, image_url=image_url, plugin_extras=plugin_extras,
+                                     params=params)
+
+    #% ---------- Group management
+    def _api_group_create(self, *, name: str, title : str = None, description: str = None,
+                          image_url: str = None, state: Union[CkanState, str] = None, params:dict=None) -> Union[CkanGroupInfo, str]:
+        """
+        Call to API group_create
+        Creates a new group (requires sysadmin user)
+        """
+        if params is None:
+            params = {}
+        params["name"] = name
+        if title is not None:
+            params["title"] = title
+        if description is not None:
+            params["description"] = description
+        if image_url is not None:
+            params["image_url"] = image_url
+        if state is not None:
+            params["state"] = str(state) if isinstance(state, CkanState) else state
+        response = self._api_action_request(f"group_create", method=RequestType.Post, json=params)
+        if response.success:
+            if params.get("return_id_only", False):
+                return response.result  # group id
+            else:
+                group_info = CkanGroupInfo.from_dict(response.result)
+                # update map
+                self.map._update_user_info(group_info.copy())
+                return group_info
+        else:
+            raise response.default_error(self)
+
+    def group_create(self, *, name: str, title : str = None, description: str = None,
+                     image_url: str = None, state: Union[CkanState, str] = None,
+                     params:dict = None, error_exists: bool = True) -> Union[CkanGroupInfo, str]:
+        """
+        Call to API group_create, checking at first if the group does not exist.
+        """
+        group_found = True
+        try:
+            _ = self.group_show(name)
+        except CkanNotFoundError:
+            group_found = False
+        if group_found and error_exists:
+            raise DuplicateNameError("group", name)
+        return self._api_group_create(name=name, title=title, description=description,
+                                      image_url=image_url, state=state,
+                                      params=params)