+++ CHANGELOG.md
@@ -7,7 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 When publishing a new release, copy the relevant section on the [Github release page](https://github.com/Mobidec/ckanapi_harvesters/releases).
 
 
-## [Unreleased] - 2026-09-30
+## [Unreleased] - 2026-09-04
+
+### Added
+
+- API calls to add group/user
+- CKAN state "pending"
 
 
 ## [0.0.38] - 2026-09-01